verify_sdk.py

try:
    from google import genai
    print("Successfully imported google.genai")
    import inspect
    print("genai members:", dir(genai))
    
    # Check Client structure
    if hasattr(genai, "Client"):
        print("Client class found.")
        # We won't instantiate with a real key, but we can check the class or a dummy instance if it doesn't validate immediately
        try:
            client = genai.Client(api_key="dummy")
            print("Client members:", dir(client))
            if hasattr(client, "responses"):
                print("client.responses exists.")
            else:
                print("client.responses DOES NOT exist.")
                
            if hasattr(client, "models"):
                print("client.models exists.")
        except Exception as e:
            print(f"Error instantiating client: {e}")
    else:
        print("Client class NOT found in google.genai")

except ImportError:
    print("Could not import google.genai")
