requirements.txt
gradio>=4.10
langgraph>=0.1.0
google-genai>=0.1.0
python-dotenv
requests
python-docx
reportlab
serpapi
beautifulsoup4
lxml
tqdm
